Refactor: move formatDate function to utils package
